// Client/projects/coi/src/app/configuration/form-builder-create/form-builder-create-interface.ts
  export interface LoadForm {
    formHeader: FormHeader
    lookupProgramElements: LookupProgramElement[]
    lookupSectionComponentType: LookupSectionComponentType[]
  }

  export interface FormHeader {
    formBuilderId: number
    formBuilderNumber: string
    versionNumber: number
    versionStatus: string
    title: string
    description: string
    isActive: string
    createTimestamp: string
    createUser: string
    updateTimestamp: string
    updateUser: string
    usages: Usage[]
    sections: Section[]
  }

  export interface Usage {
    formUsageId: number
    formBuilderId: number
    formOrderNumber: number
    moduleCode: string
    subModuleCode: string
    businessRuleId: any
    description: string
    isActive: string
    createTimestamp: string
    createUser: string
    updateTimestamp: string
    updateUser: string
  }

  export interface Section {
    sectionId: number
    formBuilderId: number
    sectionName: string
    sectionOrder: number
    businessRuleId: any
    description: string
    helpText: string
    sectionHeader: string
    sectionFooter: string
    isActive: string
    updateTimestamp: string
    updateUser: string
    sectionComponent?: SectionComponent[]
  }
// ...
  export interface LookupProgramElement {
    progElementId: number
    progElementNumber: string
    versionNumber: number
    versionStatus: string
    progElementName: string
    description: string
    isActive: string
    updateTimestamp: string
    updateUser: string
  }

  export interface LookupSectionComponentType {
    componentTypeCode: string
    description: string
    isActive: string
    updateTimestamp: string
    updateUser: string
  }

  export interface FormHeaderResponse {
    formBuilderId: number
    formBuilderNumber: string
    versionNumber: number
    versionStatus: string
    title: string
    description: string
    isActive: string
    createTimestamp: string
    createUser: string
    updateTimestamp: string
    updateUser: string
    usages: any
    sections: any
  }
// ...
  export interface formHeader {
    formBuilderId: number
    formBuilderNumber: string
    versionNumber: number
    versionStatus: string
    title: string
    description: string
    isActive: string
    createTimestamp: string
    createUser: string
    updateTimestamp: string
    updateUser: string
    usages: any
    sections: any
  }
// ...
  export class UpdateFormHeaderObject {
    formBuilderId: string;
    title: string;
    description: string;
    isActive: string;
    formDescription:string
}